requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0
alembic>=1.14.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
